requirements.txt
fastapi
uvicorn
langchain-ollama
sentence-transformers
pdfplumber
faiss-cpu
pytesseract
Pillow
python-docx
python-multipart

 